Add 'replace tags' screen, implement search
This commit adds a new page called `Replace Tag`,
and allows the user to select a tag to match.

Once selected, the form will show all of the matching
person names.

In a subsequent commit we will add a confirmation
message, and then finish the implementation to allow modification of the records.